# docs/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Speedtest Watcher API",
    "description": "Results, statistics, settings, backups and metrics of a Speedtest Watcher instance.\n\n- **Authentication:** while sign-in is off, no token is needed. While it's on, create a token under **Settings → Security → API Token** and send it as `Authorization: Bearer swt_…`. Read operations also work without a token when people who aren't signed in have read-only access.\n- **Errors** answer with `{ \"message\": \"…\" }` and a 4xx status. `401` means the request needs a valid token.\n- **Units:** timestamps are UTC ISO 8601, speeds are Mbps, ping and jitter are milliseconds, and test durations are seconds. Failed and skipped results have `-1` readings.\n\nThe guide, with examples, is at https://github.com/remotesojourner/speedtest-watcher/blob/main/docs/api.md.",
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/config": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Get the settings",
        "description": "Returns every setting the caller may see as a string, plus flags about sign-in. Read-only visitors don't get the settings marked full access only. Sign-in settings are managed on the Security tab and secrets are never returned.\n\n**Access:** read. Works for everyone while sign-in is off, and for read-only visitors, signed-in users and API tokens while it's on.",
        "responses": {
          "200": {
            "description": "The settings.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigResponse"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in is on, people who aren't signed in have no access, and the request has no valid API token or signed-in session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-access": "read"
      },
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "Change settings",
        "description": "Saves one or more settings together: either every change is saved or none is. Keys are the setting names from `GET /api/config`, and values are strings. Send `none` to unset an optional value. Sign-in settings can only be changed on the Security tab.\n\nOpen browser tabs, the scheduler and integrations pick the changes up straight away.\n\n**Access:** full. Needs sign-in to be off, a signed-in browser session, or an API token.",
        "requestBody": {
          "description": "The settings to change, as `{ \"key\": \"value\" }`.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Every change was saved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "A key is unknown or managed on the Security tab, or a value isn't allowed. Nothing was saved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in is on and the request has no valid API token or signed-in session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-access": "full"
      }
    },
...
    "/api/storage": {
      "get": {
        "tags": [
          "Storage"
        ],
        "summary": "Get storage information",
        "description": "**Access:** full. Needs sign-in to be off, a signed-in browser session, or an API token.",
        "responses": {
          "200": {
            "description": "The database size and how many results are stored.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageInfoResponse"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in is on and the request has no valid API token or signed-in session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-access": "full"
      }
    },
    "/api/storage/data": {
      "get": {
        "tags": [
          "Storage"
        ],
        "summary": "Export a data backup",
        "description": "Downloads `speedtest-watcher-data.json`: every result, and the connection monitoring history (outages, watch sessions and probe rounds). It can be imported again with `PUT /api/storage/data`.\n\n**Access:** full. Needs sign-in to be off, a signed-in browser session, or an API token.",
        "responses": {
          "200": {
            "description": "The file.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in is on and the request has no valid API token or signed-in session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-access": "full"
      },
      "put": {
        "tags": [
          "Storage"
        ],
        "summary": "Restore a data backup",
        "description": "Adds every result and every piece of connection monitoring history from a backup. Ids in the file are ignored, an entry whose timestamp is already stored is skipped, and an outage still open when the backup was made is skipped too, so importing the same file twice is safe.\n\n**Access:** full. Needs sign-in to be off, a signed-in browser session, or an API token.",
        "requestBody": {
          "description": "A backup from `GET /api/storage/data` or the Storage tab. A results-only file wraps as `{ \"speedtests\": [...] }`.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataBackup"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DataBackup"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "How much was imported and how much was skipped.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataImportResponse"
                }
              }
            }
          },
          "400": {
            "description": "The backup is empty, isn't valid, or was made by a newer version.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in is on and the request has no valid API token or signed-in session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-access": "full"
      },
      "delete": {
        "tags": [
          "Storage"
        ],
        "summary": "Delete all data",
        "description": "Deletes every result, outage, watch session and probe round. Settings, integrations and recommendations stay. This can't be undone.\n\n**Access:** full. Needs sign-in to be off, a signed-in browser session, or an API token.",
        "responses": {
          "200": {
            "description": "All data was deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in is on and the request has no valid API token or signed-in session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-access": "full"
      }
    },
    "/api/storage/tests/history": {
      "delete": {
        "tags": [
          "Storage"
        ],
        "summary": "Delete every result",
        "description": "Clears the history. Settings, integrations and recommendations stay. This can't be undone.\n\n**Access:** full. Needs sign-in to be off, a signed-in browser session, or an API token.",
        "responses": {
          "200": {
            "description": "Every result was deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in is on and the request has no valid API token or signed-in session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-access": "full"
      }
    },
    "/api/storage/config": {
      "get": {
        "tags": [
          "Storage"
        ],
        "summary": "Export a settings backup",
        "description": "Returns the settings, integrations and recommendations in the format of the Storage tab's backup file. Sign-in settings are never included.\n\n**Access:** full. Needs sign-in to be off, a signed-in browser session, or an API token.",
        "responses": {
          "200": {
            "description": "The backup.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsBackup"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in is on and the request has no valid API token or signed-in session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-access": "full"
      },
      "put": {
        "tags": [
          "Storage"
        ],
        "summary": "Restore a settings backup",
        "description": "Saves the valid settings together, adds or replaces integrations by id, and restores the recommendations when all three values are positive. Anything that isn't valid, and any sign-in setting, is skipped and counted.\n\n**Access:** full. Needs sign-in to be off, a signed-in browser session, or an API token.",
        "requestBody": {
          "description": "A backup from `GET /api/storage/config` or the Storage tab.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsBackup"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsBackup"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "What was restored and how much was skipped.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsImportResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request body isn't valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in is on and the request has no valid API token or signed-in session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-access": "full"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfigResponse": {
        "required": [
          "viewMode",
          "authActive",
          "authDisabledByEnv"
        ],
        "type": "object",
        "properties": {
          "viewMode": {
            "type": "boolean",
            "description": "The caller has read-only access."
          },
          "authActive": {
            "type": "boolean",
            "description": "Sign-in is switched on and enforced."
          },
          "authDisabledByEnv": {
            "type": "boolean",
            "description": "Sign-in is switched off by the `DISABLE_AUTH` environment variable."
          },
          "oidcClientSecretSet": {
            "type": [
              "null",
              "boolean"
            ],
            "description": "A client secret is saved for the identity provider. Left out for read-only visitors."
          },
          "apiTokenSet": {
            "type": [
              "null",
              "boolean"
            ],
            "description": "An API token exists. Left out for read-only visitors."
          },
          "ping": {
            "type": "string",
            "description": "Visible to everyone who can read the settings.",
            "default": "25"
          },
          "download": {
            "type": "string",
            "description": "Visible to everyone who can read the settings.",
            "default": "100"
          },
          "upload": {
            "type": "string",
            "description": "Visible to everyone who can read the settings.",
            "default": "50"
          },
          "maxPacketLoss": {
            "type": "string",
            "description": "Visible to everyone who can read the settings.",
            "default": "none"
          },
          "maxBufferbloat": {
            "type": "string",
            "description": "Visible to everyone who can read the settings.",
            "default": "none"
          },
          "cron": {
            "type": "string",
            "description": "Left out for read-only visitors.",
            "default": "0 * * * *"
          },
          "scheduleOffset": {
            "type": "string",
            "description": "Left out for read-only visitors.",
            "default": "true"
          },
          "unhealthyCron": {
            "type": "string",
            "description": "Left out for read-only visitors.",
            "default": "none"
          },
          "provider": {
            "type": "string",
            "description": "Visible to everyone who can read the settings.",
            "default": "none"
          },
          "interface": {
            "type": "string",
            "description": "Visible to everyone who can read the settings.",
            "default": "none"
          },
          "libreUrl": {
            "type": "string",
            "description": "Left out for read-only visitors.",
            "default": "none"
          },
          "serverMode": {
            "type": "string",
            "description": "Visible to everyone who can read the settings.",
            "default": "auto"
          },
          "serverListMode": {
            "type": "string",
            "description": "Visible to everyone who can read the settings.",
            "default": "allow"
          },
          "ooklaId": {
            "type": "string",
            "description": "Left out for read-only visitors.",
            "default": "none"
          },
          "libreId": {
            "type": "string",
            "description": "Left out for read-only visitors.",
            "default": "none"
          },
          "ooklaServerIds": {
            "type": "string",
            "description": "Left out for read-only visitors.",
            "default": "none"
          },
          "libreServerIds": {
            "type": "string",
            "description": "Left out for read-only visitors.",
            "default": "none"
          },
          "monitoringEnabled": {
            "type": "string",
            "description": "Visible to everyone who can read the settings.",
            "default": "true"
          },
          "monitoringTargets": {
            "type": "string",
            "description": "Left out for read-only visitors.",
            "default": "1.1.1.1:443,8.8.8.8:443,9.9.9.9:443"
          },
          "monitoringInterval": {
            "type": "string",
            "description": "Left out for read-only visitors.",
            "default": "15"
          },
          "monitoringRoundsDown": {
            "type": "string",
            "description": "Left out for read-only visitors.",
            "default": "3"
          },
          "monitoringRoundsUp": {
            "type": "string",
            "description": "Left out for read-only visitors.",
            "default": "2"
          },
          "monitoringTestAfterReconnect": {
            "type": "string",
            "description": "Visible to everyone who can read the settings.",
            "default": "false"
          },
          "internetCheckEnabled": {
            "type": "string",
            "description": "Visible to everyone who can read the settings.",
            "default": "true"
          },
          "internetCheckUrl": {
            "type": "string",
            "description": "Left out for read-only visitors.",
            "default": "https://icanhazip.com"
          },
          "skipIps": {
            "type": "string",
            "description": "Left out for read-only visitors.",
            "default": "none"
          },
          "chartRange": {
            "type": "string",
            "description": "Visible to everyone who can read the settings.",
            "default": "7d"
          },
          "chartBeginAtZero": {
            "type": "string",
            "description": "Visible to everyone who can read the settings.",
            "default": "false"
          },
          "dateFormat": {
            "type": "string",
            "description": "Visible to everyone who can read the settings.",
            "default": "dmy"
          },
          "retentionDays": {
            "type": "string",
            "description": "Visible to everyone who can read the settings.",
            "default": "365"
          },
          "authEnabled": {
            "type": "string",
            "description": "Changed only on the Security tab. Left out for read-only visitors.",
            "default": "false"
          },
          "visitorAccess": {
            "type": "string",
            "description": "Changed only on the Security tab. Left out for read-only visitors.",
            "default": "none"
          },
          "oidcAuthority": {
            "type": "string",
            "description": "Changed only on the Security tab. Left out for read-only visitors.",
            "default": "none"
          },
          "oidcClientId": {
            "type": "string",
            "description": "Changed only on the Security tab. Left out for read-only visitors.",
            "default": "none"
          },
          "oidcScopes": {
            "type": "string",
            "description": "Changed only on the Security tab. Left out for read-only visitors.",
            "default": "openid profile email"
          }
        },
        "description": "The saved settings, one string property per setting, plus what the caller may see and do. Unset values are `none`. Read-only visitors don't get the settings marked full access only, and secrets are never returned."
      },
...
      "DataBackup": {
        "type": "object",
        "properties": {
          "version": {
            "examples": [
              1
            ],
            "type": "integer",
            "description": "The format version. An import is rejected when this is newer than the running version knows.",
            "format": "int32"
          },
          "exported": {
            "type": "string",
            "description": "When the backup was made.",
            "format": "date-time"
          },
          "speedtests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataBackupSpeedtest"
            },
            "description": "Every result. Required: a backup with nothing in any of its four lists is rejected."
          },
          "outages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataBackupOutage"
            },
            "description": "Every connection outage."
          },
          "watchSessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataBackupWatchSession"
            },
            "description": "Every period the connection monitor was watching."
          },
          "probeRounds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataBackupProbeRound"
            },
            "description": "Every connection check behind the latency chart, kept for 30 days."
          }
        },
        "description": "A full data backup: every result, and the connection monitoring history, in the same format as the file the Storage tab exports."
      },
      "DataBackupOutage": {
        "type": "object",
        "properties": {
          "startedAt": {
            "examples": [
              "2026-09-14T08:05:00Z"
            ],
            "type": "string",
            "description": "When the outage started.",
            "format": "date-time"
          },
          "endedAt": {
            "examples": [
              "2026-09-14T08:11:00Z"
            ],
            "type": [
              "null",
              "string"
            ],
            "description": "When the outage ended. An outage still open when the backup was made, or one that ends before it starts, is skipped on import.",
            "format": "date-time"
          }
        },
        "description": "One connection outage in a data backup."
      },
...
      "DataImportResponse": {
        "required": [
          "speedtests",
          "outages",
          "watchSessions",
          "probeRounds",
          "skipped"
        ],
        "type": "object",
        "properties": {
          "speedtests": {
            "examples": [
              1204
            ],
            "type": "integer",
            "description": "Results that were stored.",
            "format": "int32"
          },
          "outages": {
            "examples": [
              6
            ],
            "type": "integer",
            "description": "Outages that were stored.",
            "format": "int32"
          },
          "watchSessions": {
            "examples": [
              40
            ],
            "type": "integer",
            "description": "Watch sessions that were stored.",
            "format": "int32"
          },
          "probeRounds": {
            "examples": [
              172800
            ],
            "type": "integer",
            "description": "Probe rounds that were stored.",
            "format": "int32"
          },
          "skipped": {
            "examples": [
              0
            ],
            "type": "integer",
            "description": "Entries that were already stored, or left out because they weren't valid, such as an outage still open when the backup was made.",
            "format": "int32"
          }
        },
        "description": "What a data backup import restored."
      },
      "ErrorResponse": {
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "What went wrong, written for people."
          }
        },
        "description": "Why a request failed. Every error the API answers carries this body."
      },
...
      "MessageResponse": {
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "What happened, written for people."
          }
        },
        "description": "Confirms that a request which changes something succeeded."
      },
...
      "SettingsBackup": {
        "type": "object",
        "properties": {
          "config": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SettingsBackupEntry"
            },
            "description": "The saved settings."
          },
          "integrations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SettingsBackupIntegration"
            },
            "description": "The configured integrations."
          },
          "recommendations": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SettingsBackupRecommendation"
              }
            ],
            "description": "The recommended targets, or `null` before 10 tests have completed."
          }
        },
        "description": "A settings backup: the settings, integrations and recommendations, in the same format as the file the Storage tab exports. Sign-in settings are never included."
      },
      "SettingsBackupEntry": {
        "type": "object",
        "properties": {
          "key": {
            "examples": [
              "cron"
            ],
            "type": "string",
            "description": "The setting's name, as in `GET /api/config`."
          },
          "value": {
            "examples": [
              "0 * * * *"
            ],
            "type": "string",
            "description": "The saved value. `none` means unset."
          }
        },
        "description": "One saved setting in a backup."
      },
...
      "SettingsImportResponse": {
        "required": [
          "settings",
          "integrations",
          "recommendations",
          "skipped"
        ],
        "type": "object",
        "properties": {
          "settings": {
            "examples": [
              21
            ],
            "type": "integer",
            "description": "Settings that were saved.",
            "format": "int32"
          },
          "integrations": {
            "examples": [
              1
            ],
            "type": "integer",
            "description": "Integrations that were added or replaced.",
            "format": "int32"
          },
          "recommendations": {
            "type": "boolean",
            "description": "Whether the recommendations were restored."
          },
          "skipped": {
            "examples": [
              0
            ],
            "type": "integer",
            "description": "Entries that were left out because they weren't valid, such as unknown settings or integration types.",
            "format": "int32"
          }
        },
        "description": "What a settings backup import restored."
      },
...
      "StorageInfoResponse": {
        "required": [
          "size",
          "testCount"
        ],
        "type": "object",
        "properties": {
          "size": {
            "examples": [
              45056
            ],
            "type": "integer",
            "description": "The size of the SQLite database, in bytes.",
            "format": "int64"
          },
          "testCount": {
            "examples": [
              41
            ],
            "type": "integer",
            "description": "How many results are stored.",
            "format": "int32"
          }
        },
        "description": "How much is stored."
      },
...
    },
    "securitySchemes": {
      "apiToken": {
        "type": "http",
        "description": "An API token (`swt_…`) from **Settings → Security → API Token**. It's needed only while sign-in is on, and read operations don't need it when people who aren't signed in have read-only access.",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "apiToken": [ ]
    },
    { }
  ],
...
}